Moved each feed provider to their own file
